Move source into src folder
This way it will be more maintainable and testable
